requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
#cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.8.0
#django-froala-editor==3.2.2
django-on-heroku==1.1.2
#django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
psycopg2-binary==2.9.9
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
